Create Exercise 6.py
The below program will draw a circle and a rectangle. Run and observe the result.
Change the radius of the circle, width and height of the rectangle using the given controls and run to observe the changes.
Your objective is to create and submit the canvas where radius of circle is 60, width of rectangle is 4 times the radius and height of the rectangle is 1 / 2 of radius.