Create Exercise 8.py
Write a Python program to calculate and display the final fees to be paid by a student.
The students are entitled to get scholarship based on the marks scored in the qualifying exam. Scholarship percentage should be considered as half of the marks scored by the student on the course fee. Apart from the course fee (after deducting the scholarship amount), the students have to pay extra service fees. Consider the course fee, marks scored by the student and extra fees amount are given. Assume that the marks are out of 100 and need not always be integer.Test your code by using the given sample inputs. Verify your code by using the 2ndsample input(highlighted) given below:
Sample Input	Expected Output
Course Fee	Marks	Extra Fees	
25000	50	1500	20250
25000	70	1500