Create Exercise 11.py
Write a python program that displays a message as follows for a given number:
a.	If it is a multiple of three, display "Zip".
b.	If it is a multiple of five, display "Zap".
c.	If it is a multiple of both three and five, display "Zoom".
d.	If it does not satisfy any of the above given conditions, display "Invalid".